Consolidate EVEN MORE logic
