+++ src/hipscat_cloudtests/__init__.py
@@ -2,4 +2,5 @@ from .example_module import greetings, meaning
 
 __all__ = ["greetings", "meaning"]
 
+from .file_checks import assert_text_file_matches
 from .temp_cloud_directory import TempCloudDirectory
